requirements.txt
crewai==0.130.0
fastapi==0.110.3
uvicorn==0.30.1
python-dotenv==1.0.1
pypdf==5.9.0
python-multipart==0.0.9
celery[redis]==5.4.0
redis==5.0.7